Installs keybase CLI binaries (#4)
TL;DR
-----

Adds Keybase CLI installation to the template

Details
-------

Updates image creation so that the resulting image has `keybase` available as a command-line (only) tool. Keybase doesn't distribute a package without the GUI, so we had to improvise. The current approach uses a Packer provisioner to run a script that downloads the Keybase package and extracts only the CLI and systemd service files.