Automates Jumpbox instance creation (#9)
TL;DR
-----

Creates an jumpbox instance after creating the image

Details
-------

Expands the automation pipeline to create a new instance
of the jumpbox after it creates an image. Right now this
image is created manually to avoid destroying a jumpbox
while it is in use.

The automation use Terraform in the same way we previously
used it to create an instance manually. Automating it as 
part of the pipeline require a few changes, most notably
using an uploaded OVA file vs. a content library to 
better model the flow as a resource in Concourse.

This change also adds a second disk to the Jumpbox as a
vSphere first-class disk to preserve home directories when 
the box is rebuilt.