Updates Jumpbox for lab based on TCE (#16)
Updates Jumpbox for lab based on TCE

TL;DR
-----

Changes the Jumpbox as needed for my rebuild with TCE

Details
-------

* Installs the TCE version of the Tanzu CLI
* Install rootless docker since the Tanzu bootstrap cluster was not
  able to work when using Podman. 

Strictly speaking the `tanzu` command can't bootstrap with rootless 
Docker either, but I wanted to make sure if I installed the Docker
Daemon it would be rootless. To bootstap a management cluster I 
need to stop rootless, start rooted, and then reverse things when
the bootstrap is done. The same is needed to delete a manaagement
cluster but I haven't done that since making this change.