Installs the MS SQL server client
TL;DR
-----

Applies lessons learned from the MS SQL image to install the client

Details
-------

Finally adds the MS SQL Server client to the Jumpbox build. I'd had
issues with it before because of EULA acceptance, but found a way
to solve for that while building the pipeline for an MS SQL Server
image. This change adds a script to install the client as part of
building the jumpbox image with Packer and includes a provisioner
that runs that script when building the image.