Corrects mount point for `/home`
TL;DR
-----

Attempts to fix error with filesystem mount for home directories

Details
-------

Noticed that the `/home` filesystem wasn't mounting and ended up
wiping it out trying to repair it. After rebuilding it still wasn't
mounting, and I noticed there might be an issue with `fstab`. This
fix updates the user data for the instance to specify the correct
partition to mount.